Centralize overarching build logic in index.ts, move context initialization up so ctx.options can be used outside of runBuild
