+++ node-src/index.ts
@@ -1,21 +1,37 @@
 import 'any-observable/register/zen';
+import Listr from 'listr';
 import readPkgUp from 'read-pkg-up';
 import { v4 as uuid } from 'uuid';
 
 import { getBranch, getCommit, getSlug, getUncommittedHash, getUserEmail } from './git/git';
+import GraphQLClient from './io/GraphQLClient';
 import HTTPClient from './io/HTTPClient';
+import NonTTYRenderer from './lib/NonTTYRenderer';
 import checkForUpdates from './lib/checkForUpdates';
 import checkPackageJson from './lib/checkPackageJson';
 import { emailHash } from './lib/emailHash';
+import { getConfiguration } from './lib/getConfiguration';
 import getEnv from './lib/getEnv';
+import getOptions from './lib/getOptions';
 import { createLogger } from './lib/log';
 import parseArgs from './lib/parseArgs';
 import { exitCodes, setExitCode } from './lib/setExitCode';
+import { rewriteErrorMessage } from './lib/utils';
 import { writeChromaticDiagnostics } from './lib/writeChromaticDiagnostics';
-import { runBuild } from './runBuild';
+import getTasks from './tasks';
 import { Context, Flags, Options } from './types';
+import { endActivity } from './ui/components/activity';
+import buildCanceled from './ui/messages/errors/buildCanceled';
+import { default as fatalError } from './ui/messages/errors/fatalError';
+import fetchError from './ui/messages/errors/fetchError';
+import graphqlError from './ui/messages/errors/graphqlError';
 import invalidPackageJson from './ui/messages/errors/invalidPackageJson';
+import missingStories from './ui/messages/errors/missingStories';
 import noPackageJson from './ui/messages/errors/noPackageJson';
+import runtimeError from './ui/messages/errors/runtimeError';
+import taskError from './ui/messages/errors/taskError';
+import intro from './ui/messages/info/intro';
+
 /**
  Make keys of `T` outside of `R` optional.
 */
@@ -56,6 +72,9 @@ export type InitialContext = Omit<
   'options'
 >;
 
+const isContext = (ctx: InitialContext): ctx is Context => 'options' in ctx;
+
+// Entry point for the CLI, GitHub Action, and Node API
 export async function run({
   argv = [],
   flags,
@@ -112,20 +131,103 @@ export async function run({
   };
 }
 
+// Entry point for testing only (typically invoked via `run` above)
 export async function runAll(ctx: InitialContext) {
-  setExitCode(ctx, exitCodes.OK);
+  ctx.log.info('');
+  ctx.log.info(intro(ctx));
 
-  ctx.http = (ctx.http as HTTPClient) || new HTTPClient(ctx);
+  try {
+    ctx.http = new HTTPClient(ctx);
+    ctx.client = new GraphQLClient(ctx, `${ctx.env.CHROMATIC_INDEX_URL}/graphql`, {
+      headers: {
+        'x-chromatic-session-id': ctx.sessionId,
+        'x-chromatic-cli-version': ctx.pkg.version,
+      },
+      retries: 3,
+    });
+    ctx.configuration = await getConfiguration(
+      ctx.extraOptions?.configFile || ctx.flags.configFile
+    );
+    (ctx as Context).options = getOptions(ctx);
+    setExitCode(ctx, exitCodes.OK);
+  } catch (e) {
+    ctx.log.info('');
+    ctx.log.error(fatalError(ctx, [e]));
+    ctx.extraOptions?.experimental_onTaskError?.(ctx, {
+      formattedError: fatalError(ctx, [e]),
+      originalError: e,
+    });
+    setExitCode(ctx, exitCodes.INVALID_OPTIONS, true);
+    return;
+  }
+
+  if (!isContext(ctx)) return;
 
   // Run these in parallel; neither should ever reject
   await Promise.all([runBuild(ctx), checkForUpdates(ctx)]);
 
-  // At this point we may or may not have options on context
   if ([0, 1].includes(ctx.exitCode) && 'options' in ctx) {
-    await checkPackageJson(ctx as Context);
+    await checkPackageJson(ctx);
   }
-  if (ctx.flags?.diagnostics || ctx.extraOptions?.diagnostics) {
-    await writeChromaticDiagnostics(ctx as Context | InitialContext);
+
+  if (ctx.options.diagnostics) {
+    await writeChromaticDiagnostics(ctx);
+  }
+}
+
+async function runBuild(ctx: Context) {
+  try {
+    try {
+      ctx.log.info('');
+      if (ctx.options.interactive) ctx.log.queue(); // queue up any log messages while Listr is running
+      const options = ctx.options.interactive ? {} : { renderer: NonTTYRenderer, log: ctx.log };
+      await new Listr(getTasks(ctx.options), options).run(ctx);
+    } catch (err) {
+      endActivity(ctx);
+      if (err.code === 'ECONNREFUSED' || err.name === 'StatusCodeError') {
+        setExitCode(ctx, exitCodes.FETCH_ERROR);
+        throw rewriteErrorMessage(err, fetchError(ctx, err));
+      }
+      if (err.name === 'GraphQLError') {
+        setExitCode(ctx, exitCodes.GRAPHQL_ERROR);
+        throw rewriteErrorMessage(err, graphqlError(ctx, err));
+      }
+      if (err.message.startsWith('Cannot run a build with no stories')) {
+        setExitCode(ctx, exitCodes.BUILD_NO_STORIES);
+        throw rewriteErrorMessage(err, missingStories(ctx));
+      }
+      if (ctx.options.experimental_abortSignal?.aborted) {
+        setExitCode(ctx, exitCodes.BUILD_WAS_CANCELED, true);
+        throw rewriteErrorMessage(err, buildCanceled());
+      }
+      throw rewriteErrorMessage(err, taskError(ctx, err));
+    } finally {
+      // Handle potential runtime errors from JSDOM
+      const { runtimeErrors, runtimeWarnings } = ctx;
+      if ((runtimeErrors && runtimeErrors.length) || (runtimeWarnings && runtimeWarnings.length)) {
+        ctx.log.info('');
+        ctx.log.error(runtimeError(ctx));
+      }
+
+      ctx.log.flush();
+    }
+  } catch (error) {
+    const errors = [].concat(error); // GraphQLClient might throw an array of errors
+    const formattedError = fatalError(ctx, errors);
+
+    ctx.options.experimental_onTaskError?.(ctx, {
+      formattedError,
+      originalError: errors[0],
+    });
+
+    if (!ctx.userError) {
+      ctx.log.info('');
+      ctx.log.error(formattedError);
+    }
+
+    if (!ctx.exitCode) {
+      setExitCode(ctx, exitCodes.UNKNOWN_ERROR);
+    }
   }
 }
